requirements.txt
undetected-chromedriver>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=13.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
textblob>=0.17.0
python-dotenv>=1.0.0
psutil>=5.9.0
